requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.6.1=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.10.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
blinker=1.4=py38haa95532_0
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2022.4.26=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.5.18.1=py38haa95532_0
cffi=1.15.0=py38h2bbff1b_1
cfitsio=3.470=he774522_6
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=py38haa95532_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.8=py38h71e12ea_0
cudatoolkit=11.1.1=heb2d755_9
cvxopt=1.2.6=pypi_0
cvxpy=1.1.11=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.22=pypi_0
cytoolz=0.11.0=py38he774522_0
dask-core=2022.2.1=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
ecos=2.0.7.post1=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.2.0=py38h2bbff1b_0
fsspec=2022.2.0=pyhd3eb1b0_0
giflib=5.2.1=h62dcd97_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.42.0=py38hc60d5dd_0
h5py=3.6.0=py38h3de5c98_0
hdf5=1.10.6=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py38ha1f97ea_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.11.4=pypi_0
importlib-resources=5.7.1=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.13.0=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9b=hb83a4c4_2
json5=0.9.8=pypi_0
jsonschema=4.6.0=pypi_0
jupyter-client=7.3.1=pypi_0
jupyter-core=4.10.0=pypi_0
jupyter-packaging=0.12.1=pypi_0
jupyter-server=1.17.0=pypi_0
jupyterlab=3.4.2=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.14.0=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.3.2=py38hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.8=h2bbff1b_5
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
libzopfli=1.0.3=ha925a31_0
locket=0.2.1=py38haa95532_2
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py38haa95532_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py38haa95532_1
matplotlib-base=3.5.1=py38hd77b12b_1
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=5.1.0=py38h2bbff1b_2
munkres=1.1.4=py_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.4=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
nest-asyncio=1.5.5=pypi_0
networkx=2.7.1=pyhd3eb1b0_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
notebook=6.4.11=pypi_0
notebook-shim=0.1.0=pypi_0
numpy=1.21.5=py38h7a0a035_2
numpy-base=1.21.5=py38hca35cd5_2
oauthlib=3.2.0=pyhd3eb1b0_0
open3d=0.15.1=pypi_0
opencv-python=4.6.0.66=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1p=h2bbff1b_0
osqp=0.6.2.post0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pyhd3eb1b0_1
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py38hdc2b20a_0
pip=21.2.2=py38haa95532_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=3.20.1=py38hd77b12b_0
psutil=5.9.1=pypi_0
pure-eval=0.2.2=pypi_0
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydensecrf=1.0rc2=pypi_0
pygments=2.12.0=pypi_0
pyjwt=2.1.0=py38haa95532_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38hd77b12b_6
pyreadline=2.1=py38_1
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.9.1=py3.8_cuda11.1_cudnn8_0
pytz=2022.1=pypi_0
pywavelets=1.3.0=py38h2bbff1b_0
pywin32=304=pypi_0
pywinpty=2.0.5=pypi_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=23.0.0=pypi_0
qdldl=0.1.5.post0=pypi_0
qpsolvers=1.7.0=pypi_0
qt=5.9.7=vc14h73c81de_0
quadprog=0.1.8=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.19.2=py38hf11a4ad_0
scipy=1.7.3=py38h0a974cb_0
scs=2.1.4=pypi_0
send2trash=1.8.0=pypi_0
setuptools=59.5.0=pypi_0
sip=4.19.13=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.2=h2bbff1b_0
stack-data=0.2.0=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py38haa95532_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pyhd3eb1b0_0
terminado=0.15.0=pypi_0
thop=0.1.0-2207010342=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
tinycss2=1.1.1=pypi_0
tk=8.6.11=h2bbff1b_0
tomlkit=0.11.0=pypi_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.9.1=py38
torchsummary=1.5.1=pypi_0
torchvision=0.10.1=py38_cu111
tornado=6.1=py38h2bbff1b_0
tqdm=4.64.0=py38haa95532_0
traitlets=5.2.2.post1=pypi_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py38haa95532_0
vc=14.2=h21ff451_1
visualize=0.5.0=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h2bbff1b_0
zfp=0.5.5=hd77b12b_6
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
